+++ student_morozov_nikita/src/test/resources/schema.sql
@@ -35,8 +35,8 @@ CREATE UNIQUE INDEX ix_country_default_day_rate_country_ic ON country_default_da
 CREATE TABLE IF NOT EXISTS age_coefficient
 (
     id          BIGINT         NOT NULL AUTO_INCREMENT,
-    age_from    BIGINT         NOT NULL,
-    age_to      BIGINT         NOT NULL,
+    age_from    INTEGER         NOT NULL,
+    age_to      INTEGER         NOT NULL,
     coefficient DECIMAL(10, 2) NOT NULL,
     PRIMARY KEY (id)
 );
